Why IdentityT is used in optparse-applicative examples
This gets confusing for newcomers since it creates clutter that they are
trying to avoid by using library, as this one.

Reason for using IdentityT, in examples involving optparse-applicative,
is that this library is trying, hard, to avoid orphan instances.

In other words IdentityT is used as something like "WrappedApplicative"
to be able to operate on any Applicative Functor without the need to
have concrete instances for all the specific data types.